Important tools Update Notice
Due to issues with the low-level C implementation, we have temporarily replaced it with a Python script. While the system is currently operational, as briefly described in the README document, this vulnerability exhibits highly unstable behavior during testing. Triggering the vulnerability may require sending a large number of data packets, and the behavior may vary each time. As a result, remote code execution might be challenging, but we encourage you to try.

For detailed information about this update, we recommend carefully reviewing the code differences. We sincerely apologize for the slow pace of updates.

Thank you for your understanding and support.